# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Incidents, service requests, workflow, assignment, comments and audit for IT service management. Errors are RFC 9457 problem details (`ApiProblem`); clients should branch on `code`.",
    "title" : "Enterprise Service Management API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Same origin as the web app"
  } ],
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
  "paths" : {
...
    "/api/admin/teams" : {
      "get" : {
        "operationId" : "listAllTeams",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TeamResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "List all teams, including inactive ones",
        "tags" : [ "Admin: teams" ]
      },
      "post" : {
        "operationId" : "createTeam",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamResponse"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Create a team",
        "tags" : [ "Admin: teams" ]
      }
    },
    "/api/admin/teams/{id}" : {
      "patch" : {
        "operationId" : "updateTeam",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Rename, describe, activate or deactivate a team (requires current version)",
        "tags" : [ "Admin: teams" ]
      }
    },
    "/api/admin/teams/{id}/members/{userId}" : {
      "delete" : {
        "operationId" : "removeTeamMember",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Remove a user from a team (idempotent)",
        "tags" : [ "Admin: teams" ]
      },
      "put" : {
        "operationId" : "addTeamMember",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Add a user to a team (idempotent)",
        "tags" : [ "Admin: teams" ]
      }
    },
...
    "/api/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Sign in. Returns an access token; sets the refresh-token cookie",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/auth/logout" : {
      "post" : {
        "operationId" : "logout",
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "End this session: revokes the refresh-token family and clears the cookie",
        "tags" : [ "Authentication" ]
      }
    },
    "/api/auth/refresh" : {
      "post" : {
        "operationId" : "refreshSession",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Exchange the refresh-token cookie for a new access token (rotates the cookie)",
        "tags" : [ "Authentication" ]
      }
    },
...
    "/api/teams" : {
      "get" : {
        "operationId" : "listTeams",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TeamResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "List active teams",
        "tags" : [ "Teams" ]
      }
    },
    "/api/teams/{id}" : {
      "get" : {
        "operationId" : "getTeam",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Get an active team",
        "tags" : [ "Teams" ]
      }
    },
...
    "/api/users/me" : {
      "get" : {
        "operationId" : "getMyProfile",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyProfileResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Profile of the signed-in user, including their teams",
        "tags" : [ "Current user" ]
      }
    },
    "/api/users/me/password" : {
      "post" : {
        "operationId" : "changeMyPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Change own password. Signs out all sessions; the client must sign in again",
        "tags" : [ "Current user" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ApiProblem" : {
        "type" : "object",
        "description" : "RFC 9457 problem details, returned for every error",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Stable machine-readable code: branch on this, never on detail",
            "enum" : [ "VALIDATION_FAILED", "MALFORMED_REQUEST", "AUTHENTICATION_REQUIRED", "ACCESS_DENIED", "RESOURCE_NOT_FOUND", "METHOD_NOT_ALLOWED", "UNSUPPORTED_MEDIA_TYPE", "CONCURRENT_MODIFICATION", "INVALID_SORT_PROPERTY", "INVALID_CREDENTIALS", "REFRESH_TOKEN_INVALID", "INVALID_ORIGIN", "CURRENT_PASSWORD_INCORRECT", "EMAIL_ALREADY_EXISTS", "LAST_ADMIN_REQUIRED", "TEAM_NAME_ALREADY_EXISTS", "TEAM_INACTIVE", "USER_NOT_ELIGIBLE_FOR_TEAM", "CATEGORY_CODE_ALREADY_EXISTS", "CATEGORY_TEAM_REQUIRED", "INVALID_CATEGORY", "INVALID_STATUS_TRANSITION", "TRANSITION_NOT_PERMITTED", "TRANSITION_REQUIREMENT_MISSING", "TICKET_NOT_ASSIGNED", "ASSIGNMENT_NOT_PERMITTED", "ASSIGNEE_NOT_IN_TEAM", "TICKET_NOT_ASSIGNABLE", "MEMBER_HAS_OPEN_TICKETS", "COMMENT_NOT_PERMITTED", "TICKET_CLOSED", "REQUEST_FAILED", "INTERNAL_ERROR" ]
          },
          "correlationId" : {
            "type" : [ "string", "null" ],
            "description" : "Matches the X-Request-Id response header and server log lines"
          },
          "detail" : {
            "type" : [ "string", "null" ],
            "description" : "Human-readable explanation; never contains internal details"
          },
          "fieldErrors" : {
            "type" : [ "array", "null" ],
            "description" : "Present for VALIDATION_FAILED",
            "items" : {
              "$ref" : "#/components/schemas/FieldErrorDetail"
            }
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "description" : "Request path",
            "example" : "/api/tickets/0199b2c4-6a1e-7c3e-9f00-1a2b3c4d5e6f/transitions"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 409
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "title" : {
            "type" : "string",
            "example" : "Conflict"
          },
          "type" : {
            "type" : "string",
            "description" : "Problem type URI",
            "example" : "about:blank"
          }
        },
        "required" : [ "code", "correlationId", "detail", "fieldErrors", "instance", "status", "timestamp", "title", "type" ]
      },
...
      "AuthenticatedUser" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "REQUESTER", "AGENT", "TEAM_LEAD", "ADMIN" ]
          }
        },
        "required" : [ "displayName", "email", "id", "role" ]
      },
...
      "ChangePasswordRequest" : {
        "type" : "object",
        "properties" : {
          "currentPassword" : {
            "type" : "string",
            "maxLength" : 256,
            "minLength" : 0
          },
          "newPassword" : {
            "type" : "string",
            "maxLength" : 72,
            "minLength" : 12
          }
        },
        "required" : [ "currentPassword", "newPassword" ]
      },
...
      "CreateTeamRequest" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          }
        },
        "required" : [ "name" ]
      },
...
      "FieldErrorDetail" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "field", "message" ]
      },
...
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "maxLength" : 254,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "maxLength" : 256,
            "minLength" : 0
          }
        },
        "required" : [ "email", "password" ]
      },
      "MyProfileResponse" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "REQUESTER", "AGENT", "TEAM_LEAD", "ADMIN" ]
          },
          "teams" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TeamSummary"
            }
          }
        },
        "required" : [ "displayName", "email", "id", "role", "teams" ]
      },
...
      "TeamResponse" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "active", "createdAt", "description", "id", "name", "updatedAt", "version" ]
      },
      "TeamSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
...
      "TokenResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "expiresIn" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tokenType" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/AuthenticatedUser"
          }
        },
        "required" : [ "accessToken", "expiresIn", "tokenType", "user" ]
      },
...
      "UpdateTeamRequest" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "description" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0,
            "pattern" : ".*\\S.*"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "version" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}